backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.47
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==1.5.3
openpyxl==3.1.2
python-multipart
bcrypt==3.2.2